+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Minor: Include discord user profile in notification metadata. (#226)
 - Minor: Add setting to allow death notifications during safe activities. (#227)
 - Minor: Allow player name in Discord notifications to link to CollectionLog.net profile. (#224)
+- Bugfix: Ensure lost items metadata is empty upon safe deaths in TzHaar Fight Cave or Inferno. (#230)
 
 ## 1.4.1